4/ Deferred Resolvers and Fetchers - (1/2)
There are two problems. At first, it will not be compiled. If you want
to compile properly, uncomment `HasId` part. Will be solved in next
part. After uncommented, it comes with an error about `deferOpt`. It
could be solved next half but I am not sure because I dont read that
yet.