7/ Preparing database
Add `postedBy` column to Link table and set it as foreign key of Id in
User table. `userId` and `linkId` of Vote table is set as foreign key of
Id in User table and Id in Link table. Aranged to define tables and run
queries for tables' heirarchy. User should be before on Link and Vote,
Link should be before on Vote.